HACK: efi: Disable overlap check
The current code checks that no partitions overlap with the GPT partition
table using the offset of the first LBA usable for that partition. This
works fine, unless you want to create partitions for something that is
between the partition table and the partition descriptors, for example a
bootloader.

Disable it for now

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>